6/13 기능수정 1. Tree를 만들어서 RootNode가 생성될 때 SharedData가 같이 만들어져 RootNode에 변수에 할당시킴 => project탭에 Tree의 SO아래로 들어가기 때문에 찾기도 편해지고 Tree따로 SharedData따로 만들지 않아도 됨 2. 프리팹으로 존재하는 객체는 BehaviorTreeEditor에 Variables탭에서 값을 정해주도록함 3. 씬에 존재하는 객체는 해당 객체의 인스펙터에서 Tree를 할당해 값을 정해주도록 함

+++ Assets/BehaviorTreeTool/Scripts/Conditions/Wait.cs
@@ -6,7 +6,7 @@ namespace BehaviorTreeTool.Scripts.Conditions
     public class Wait : ConditionNode
     {
         private float _startTime;
-        [SerializeField] private int duration = 1;
+        [SerializeField] private float duration = 1;
 
         private string _remainingTimeText;
